backend/requirements.txt
Django==5.2.*
djangorestframework==3.15.*
django-cors-headers==4.4.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
celery==5.4.*
redis==5.0.*
djangorestframework-simplejwt==5.3.*
pydantic==2.8.*
gunicorn==22.0.*
uvicorn[standard]==0.30.*
# dev
pytest==8.*
pytest-django==4.*
ruff==0.5.*
